demo/requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
characteristic==14.3.0
chardet==3.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==4.0.3
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
falcon==2.0.0
Flask==2.0.2
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
jasmin==0.10.9
Jinja2==3.0.3
klein==21.8.0
kombu==5.2.3
lockfile==0.12.2
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
prompt-toolkit==3.0.24
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyHamcrest==2.0.3
PyJWT==2.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-messaging==0.5.13
python-mimeparse==1.6.0
pytz==2021.3
redis==3.4.1
requests==2.23.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
smpp.pdu3==0.6
smpp.twisted3==0.6
sqlparse==0.4.2
tabulate==0.8.9
tk==0.1.0
treq==20.3.0
Tubes==0.2.0
Twisted==20.3.0
txAMQP3==0.9.1
txredisapi==1.4.7
uritemplate==4.1.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
zope.interface==5.4.0
